Добавлен репозиторий тем. Подправил репозиторий авторов.

+++ src/main/java/ru/siv/articles/repository/TheAuthorsRepository.java
@@ -8,7 +8,7 @@ import ru.siv.articles.model.TheAuthors;
 import java.util.List;
 
 @Repository
-public interface TheAuthorsRepository extends CrudRepository<TheAuthors, Integer> {
+public interface TheAuthorsRepository extends CrudRepository<TheAuthors, Long> {
   TheAuthors findByNameIgnoreCase(String name);
   @Query("SELECT ta FROM TheAuthors ta ORDER BY ta.name")
   List<TheAuthors> filterAllOrderByName();
